Split Inventory.Action to its own module
